Moved the Ship outsde of the Ship Console Controller into the Controllers root for accessibility
