+++ StarshipAPI/Models/StarshipContext.cs
@@ -22,7 +22,7 @@ namespace StarshipAPI.Models
 
         public DbSet<ModuleUnitRoom> ModuleUnitRoom { get; set; }
 
-        public DbSet<Operator> Operator { get; set; }
+        public DbSet<SectorManager> SectorManager { get; set; }
         public DbSet<ShipModuleUnitRoom> ShipModuleUnitRoom { get; set; }
     }
 }
